Refactor applicantInformation page to use veteranId (#5528)
* Refactor applicantInformation page to use veteranId

* Rename veteranId as personId
