+++ src/js/edu-benefits/pages/applicantInformation.js
@@ -2,9 +2,9 @@ import _ from 'lodash/fp';
 
 import currentOrPastDateUI from '../../common/schemaform/definitions/currentOrPastDate';
 import fullNameUI from '../../common/schemaform/definitions/fullName';
-import ssnUI from '../../common/schemaform/definitions/ssn';
 
 import { relationshipLabels, genderLabels } from '../utils/labels';
+import * as personId from '../definitions/personId';
 
 
 const defaults = (prefix) => {
@@ -39,7 +39,6 @@ export default function applicantInformation(schema, options) {
   const prefix = (options && options.isVeteran) ? 'veteran' : 'relative';
   const mergedOptions = _.assign(defaults(prefix), options);
   const { fields, required, labels } = mergedOptions;
-  const fileNumberProp = prefix === 'relative' ? 'relativeVaFileNumber' : 'vaFileNumber';
 
   const possibleProperties = _.assign(schema.properties, {
     'view:noSSN': {
@@ -51,25 +50,9 @@ export default function applicantInformation(schema, options) {
     path: 'applicant/information',
     title: 'Applicant information',
     initialData: {},
-    uiSchema: {
+    uiSchema: _.assign({
       'ui:order': fields,
       [`${prefix}FullName`]: fullNameUI,
-      [`${prefix}SocialSecurityNumber`]: _.assign(ssnUI, {
-        'ui:required': (formData) => !_.get('view:noSSN', formData)
-      }),
-      'view:noSSN': {
-        'ui:title': 'I don’t have a Social Security number',
-      },
-      [fileNumberProp]: {
-        'ui:required': (formData) => !!_.get('view:noSSN', formData),
-        'ui:title': 'File number',
-        'ui:errorMessages': {
-          pattern: 'File number must be 8 digits'
-        },
-        'ui:options': {
-          expandUnder: 'view:noSSN'
-        }
-      },
       [`${prefix}DateOfBirth`]: _.assign(currentOrPastDateUI('Date of birth'),
         {
           'ui:errorMessages': {
@@ -92,7 +75,7 @@ export default function applicantInformation(schema, options) {
           labels: labels.relationship || relationshipLabels
         }
       }
-    },
+    }, personId.uiSchema(prefix, 'view:noSSN')),
     schema: {
       type: 'object',
       definitions: _.pick([
